+++ devServer.js
@@ -1,0 +1,32 @@
+var path = require('path');
+var express = require('express');
+var webpack = require('webpack');
+var config = require('./webpack.config.dev');
+
+var app = express();
+
+var compiler = webpack(config);
+
+app.use(require('webpack-dev-middleware')(compiler, {
+  publicPath: config.output.publicPath,
+  contentBase: 'src',
+  stats: {
+    colors: true,
+    hash: false,
+    timings: true,
+    chunks: false,
+    chunkModules: false,
+    modules: false
+  }
+}));
+
+app.use(require('webpack-hot-middleware')(compiler));
+
+app.use(express.static(__dirname + '/dist'));
+
+app.listen(config._hotPort, 'localhost', function(err) {
+  if (err) {
+    console.log(err);
+  }
+  console.info("==> 🌎 Listening on port %s. Open up http://localhost:%s/ in your browser.", config._hotPort, config._hotPort);
+});