+++ prodServer.js
@@ -1,0 +1,16 @@
+var path = require('path');
+var express = require('express');
+var app = express();
+
+app.use(express.static(__dirname + '/dist'));
+
+app.get('*', function(req, res) {
+  res.sendFile(path.join(__dirname, 'dist/index.html'));
+});
+
+app.listen(8000, 'localhost', function(err) {
+  if (err) {
+    console.log(err);
+  }
+  console.info("==> 🌎 Listening on port %s. Open up http://localhost:%s/ in your browser.", 8000, 8000);
+});