+++ src/components/App.js
@@ -1,18 +1,15 @@
 import React, {PropTypes} from 'react';
-import {RouteHandler} from 'react-router';
 
 const App = React.createClass({
 
-  propTypes: {
-    pathname: PropTypes.string.isRequired
+  propTypes: typeof __DEV__ && {
+    children: PropTypes.object.isRequired
   },
 
   render() {
-    const {pathname} = this.props;
-
     return (
       <div>
-        <RouteHandler {...this.props} key={pathname} />
+        {this.props.children}
       </div>
     );
   }