+++ src/components/Counter.js
@@ -1,15 +1,15 @@
 import React, {PropTypes} from 'react';
-import './Counter.css';
+import style from './Counter.styl';
 
 const Counter = React.createClass({
 
-  propTypes: {
+  propTypes: __DEV__ && {
     count: PropTypes.number.isRequired,
     onIncrement: PropTypes.func.isRequired
   },
 
   increment() {
-    // maybe I do something here
+    // maybe I'd like to do something here
     // for the moment I just call the callback passed from the parent
     return this.props.onIncrement();
   },
@@ -18,15 +18,12 @@ const Counter = React.createClass({
     const {count} = this.props;
 
     return (
-      <div className="Counter">
-        <h1 className="h1 h0-responsive caps mt4 mb0 regular">Count: {count}</h1>
-        <p className="h3">Click the button to increment the counter</p>
-        <a
-          onClick={this.increment}
-          className="h3 btn btn-primary mb4"
-        >
+      <div className={style.counter}>
+        <h1>Count: {count}</h1>
+        <p>Click the button to increment the counter</p>
+        <button className="p1" onClick={this.increment}>
           Increment
-        </a>
+        </button>
       </div>
     );
   }