+++ src/main.js
@@ -1,7 +1,10 @@
 import 'babel/polyfill';
-import 'basscss/css/basscss.css';
+import 'normalize.css';
+import './globals.styl';
 import React from 'react';
+import ReactDOM from 'react-dom';
 import Router from 'react-router';
+import createBrowserHistory from 'history/lib/createBrowserHistory';
 import routes from './routes';
 import attachFastClick from 'fastclick';
 
@@ -11,12 +14,10 @@ attachFastClick.attach(document.body);
 // Expose globally
 window.React = React;
 
-const router = Router.create({
-  routes: routes,
-  location: Router.HashLocation // Router.HistoryLocation
-});
+const history = createBrowserHistory();
 
-router.run((Handler, state) => {
-  React.render(<Handler {...state} />,
-    document.getElementById('root'));
-});
+ReactDOM.render(
+  <Router
+    children={routes}
+    history={history} />,
+  document.getElementById('root'));