+++ test/test-init.js
@@ -1,4 +1,7 @@
-import chai, { expect } from 'chai';
+import chai, {
+  expect
+}
+from 'chai';
 import sinon from 'sinon';
 import sinonChai from 'sinon-chai';
 
@@ -6,8 +9,12 @@ chai.use(sinonChai);
 
 global.expect = expect;
 global.sinon = sinon;
+global.__DEV__ = true;
 
 // ensure requiring css not to throw
-require.extensions['.css'] = function () {
+require.extensions['.css'] = function() {
+  return null;
+};
+require.extensions['.styl'] = function() {
   return null;
 };